feat: Add new DTOs and update .gitattributes for consistent line endings
- Introduced multiple new DTO classes for admin functionalities, including AdminOrderDTO, AdminProductDTO, AdminSellerDTO, AdminStatsDTO, AdminUserDTO, and others to enhance data handling.
- Updated .gitattributes to enforce consistent line endings for Java, properties, XML, YAML, and Markdown files, ensuring better cross-platform compatibility.

+++ e-commerce-backend/src/main/java/com/example/repositories/SubCategoryRelationRepository.java
@@ -11,7 +11,7 @@ import com.example.models.SubCategoryRelation;
 import java.util.List;
 
 @Repository
-public interface SubCategoryRelationRepository extends JpaRepository<SubCategoryRelation, Integer> {
+public interface SubCategoryRelationRepository extends JpaRepository<SubCategoryRelation, Long> {
     
     // Find all direct subcategories (depth=1) of a given category
     List<SubCategoryRelation> findByAncestorAndDepth(Category ancestor, int depth);
